Add sphinx-tabs support
sphinx-tabs support was added with the help of user-macros that are
required to be installed on the confluence instance being published to.
See README.rst update in this commit for details of the user-macros.
The implementation is a little 'hacky' but does the job. It requires
that sphinx-tabs does not recognize 'confluence' as a supported builder
type as the structure outputted in the 'unregonized builder' mode is a
little easier to handle and is also the default state of sphinx-tabs. It
is important to NOT set 'sphinx_tabs_valid_builders' configuration
option to include 'confluence'. This implementation is mostly
isolated in the 'visit/depart_container' translatort.py methods with
only a small change in 'visit/depart_paragraph' to suppress the 'p' tag.