Set height of tabs after img tags have loaded src
The current user-macros for confluence tabs sometimes would not set the
proper height of the tab-container. This mainly would happen when the
content within the tabs had to be loaded (i.e. images). The macro was
updated to check for completion of all images before setting the height.
This is not perfect but in most scenarios this will work just fine. The
macro will check every 50 milliseconds for the first 500ms and then
500ms after that. At this point it works so I am going to leave the
implementation as is but since there is javascript being used already I
think I could look to move to a 100% javascript implementation instead
of relying on this radio button method. The would provide the benefit of
not having to worry about waiting for content to load or not as the tab
can be a proper container that can just expand/contract as tabs are
selected.